python/requirements.txt
pytest
psycopg2
flask
flask-sqlalchemy
geoalchemy2
requests
pillow
redis
